requirements.txt
Babel==2.1.1
blinker==1.4
coverage==4.0.3
decorator==4.0.6
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
guess-language==0.2
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
lxml==3.5.0
MarkupSafe==0.23
mecab-python==0.996
pbr==1.8.1
python-openid==2.2.5
pytz==2015.7
rauth==0.7.2
requests==2.9.1
six==1.10.0
speaklater==1.3
SQLAlchemy==1.0.10
sqlalchemy-migrate==0.10.0
sqlparse==0.1.18
Tempita==0.5.2
Werkzeug==0.11.3
Whoosh==2.7.0
WTForms==2.1
